==> notebooks/02_Initial_Data_Processing_And_Feature_Engineering.ipynb <==
<center><h1><font size=6> Initial Data Processing and Feature Engineering </h1></center>

This notebook takes the raw EPL match data that was scraped in Notebook 01 and processes the data to make sure it is in a clean and consistent format, including:
* Cleaning up column names and descriptions
* Cleaning column types
* Ensuring team names are consistent across the dataset

It then takes the processed data and calculates some simple features that I thought were useful based on my domain knowledge of the EPL, including:

* **Days since last game and number of games in last 21 days**: this is an important indicator of fatigue.
* **Points, total points and goal difference in the current season**. This is a good indicator of how well a team is doing in the current season.
* **League position**: this is an important indicator of form. 
* **Stats from previous season**: this is a good indicator of how good a team is.
* **Last head to head results**: the most recent result and form of the team against the same opponent (and at the same venue)
* **Promoted dummy variable**: this could be an important indicator and is a useful variable to capture as promoted teams are likely to have less available data on historic form so will need to be treated slightly differently.

More machine learning specific processing and feature engineering is done after the Exploratory Data Analysis (EDA) phase, but I wanted to do this after I split the data into training and testing sets to prevent any snooping bias.

### Load libraries and setup notebook configuration

# import packages
import pandas as pd 
import numpy as np
import os
from pathlib import Path
import warnings
import datetime


# set pandas configurations
pd.set_option("display.precision", 2) # display to 1 decimpal place
pd.set_option("display.max.columns", None) # display all columns so we can view the whole dataset
pd.set_option('display.float_format', '{:.2f}'.format) # Disable scientific notation for pandas
warnings.filterwarnings("ignore", category=pd.errors.SettingWithCopyWarning) # Disable setting with copy warnings


# set directories
os.chdir('..') # change current working directory to the parent directory to help access files/directories at a higher level
DATAPATH = Path(r'data') # set data path


# import from source directory
from src import constants

### Load data from local data file

# load EPL match data
matches_raw = pd.read_csv(f"{DATAPATH}/raw/matches_long_raw.csv")

matches_raw.head(5)

<center><h1><font size=6> Data Processing </h1></center>

### Clean data

# collect only relevant columns (some we drop because of inconsistent reporting (e.g., only being reported in very early seasons))
matches_clean = matches_raw[['Date', 'Time', 'Comp', 'Round', 'Day', 'Venue', 'Result', 'GF', 'GA', 'Opponent', 'xG', 'xGA', 'Poss', 'Attendance', 'Captain',
                             'Formation', 'Referee', 'Match Report', 'Notes', 'season', 'team','date_downloaded']]


# clean column names
matches_clean.columns = map(str.lower, matches_clean.columns) # make columns lower case
matches_clean.columns = matches_clean.columns.str.replace(' ', '_') # replace spaces in column names

# Convert date into datetime object
matches_clean['date'] = pd.to_datetime(matches_clean['date'])

# Convert time into time object 
matches_clean['time'] = pd.to_datetime(matches_clean['time'], format='%H:%M').dt.time

# Convert date downloaded into datetime object 
matches_clean['date_downloaded'] = pd.to_datetime(matches_clean['date_downloaded'])

# Convert goals for to numeric 
matches_clean.loc[:, 'gf'] = pd.to_numeric(matches_clean['gf'], errors='coerce', downcast='integer')

# Convert goals against to numeric 
matches_clean.loc[:, 'ga'] = pd.to_numeric(matches_clean['ga'], errors='coerce', downcast='integer')

# convert matchweek into a single number
matches_clean['round'] = matches_clean['round'].str.extract(r'Matchweek (\d+)', expand=False).fillna(matches_clean['round'])

# Remove any rows without a date, team, or opponent
columns_to_check = ['team', 'opponent', 'date']
matches_clean = matches_clean.dropna(subset=columns_to_check)

### Create a unique game ID for each game

# create an ID based on the date, venue, team and opponent
matches_clean['unique_match_id'] = (matches_clean['date'].dt.strftime("%Y%m%d")
                                    + pd.factorize(matches_clean['venue'])[0].astype(str)
                                    + pd.factorize(matches_clean['team'])[0].astype(str) 
                                    + pd.factorize(matches_clean['opponent'])[0].astype(str))
matches_clean['unique_match_id'] = matches_clean['unique_match_id'].astype('int64')
matches_clean = matches_clean[['unique_match_id'] + [col for col in matches_clean.columns if col != 'unique_match_id']] # make ID first column

# Remove rows with duplicate values in 'unique_match_id' column
matches_clean.drop_duplicates(subset='unique_match_id', inplace=True)

### Cleaning team names

One issue we have with the data in its raw format is that the team names are different in the 'team' and 'opponent' column, which makes it dificult to map across data in future. So let's create a mapping to ensure all our team names are consistent.

# collect team names
pl_matches = matches_clean[matches_clean['comp'] == "Premier League"] # filter to just contain PL opponents otherwise we get european and cup game opponents as well

teams = np.sort(pl_matches.team.unique())
opponents = np.sort(pl_matches.opponent.unique())


# find the index points within each list where the names are not equal
unequal_indexes = [i for i, (x, y) in enumerate(zip(teams, opponents)) if x != y]

for index in unequal_indexes:
    print(f"Team name [{index}]: {teams[index]}, Opponenent name [{index}]: {opponents[index]}")

# map the opponent team names to the team names to ensure consistency
team_mapping = dict(zip(opponents, teams)) # create a dictionary mapping each unique opponent name to the team name
matches_clean['opponent'] = matches_clean['opponent'].map(lambda x: team_mapping.get(x, x)) # map to new names but leave original name if no mapping found (e.g., champions league games)

<center><h1><font size=6> Basic Feature Engineering </h1></center>

matches = matches_clean.copy().sort_values(['team', 'season', 'date']).reset_index(drop=True)

### Days since last game

# calculate days since last game
matches['days_since_last_game'] = matches.groupby(['team', 'season'])['date'].diff().dt.days

### Amount of games played in last 21 days

# calculate count of games in previous 21 days up to current day for each team
matches['games_played_last_21_days'] = matches.set_index('date')\
                .groupby('team', sort=False)['unique_match_id']\
                .rolling('21d', closed='left').count().tolist()

### Points, total points and goal difference (per game)

# calculate the number of points in each game based on the result
# Function to calculate points from a game
def calculate_points(row):
    if row['result'] == 'W':
        return 3
    elif row['result'] == 'D':
        return 1
    else:
        return 0

# calculate points
matches['points'] = matches.apply(calculate_points, axis=1)


# define a function to calculate the cumulative amount of a certain variable over the course of a PL season (up to but not including the game of the row)
def calculate_cumulative_pl_value(data, new_column_name, column):
    data[new_column_name] = data[data['comp'] == 'Premier League'].groupby(['team', 'season'])[column].transform(lambda x: x.shift().expanding().mean())
    
    # Identify the transition points between seasons
    season_transition_mask = data['round'] == '1'
    season_start_mask = season_transition_mask & (data['comp'] == 'Premier League')
    
    # Reset the cumulative values at the start of each new season
    data.loc[season_start_mask, new_column_name] = 0
    
    return data

    
# calculate total cumulative points, goals for and goals against
matches = calculate_cumulative_pl_value(data=matches, new_column_name='pl_total_points', column='points')
matches = calculate_cumulative_pl_value(data=matches, new_column_name='pl_total_gf', column='gf')
matches = calculate_cumulative_pl_value(data=matches, new_column_name='pl_total_ga', column='ga')


# calculate total cumulative goal difference
matches['pl_total_goal_diff'] = matches['pl_total_gf'] - matches['pl_total_ga']

columns_to_convert = ['pl_total_points', 'pl_total_gf', 'pl_total_ga', 'pl_total_goal_diff']
matches[columns_to_convert] = matches[columns_to_convert].astype(float)

### Relative position in the PL table at each matchweek

def calculate_league_position(data):
    # Create a copy of the dataframe to avoid modifying the original one
    temp_df = data[data['comp'] == 'Premier League'].copy()

    # Convert 'round' to integer matchweek number
    temp_df['matchweek'] = temp_df['round']

    # Create a new dataframe to hold the ranks
    ranks = pd.DataFrame()

    # For each season and each matchweek, sort teams by total points, then by goal difference and finally by team name
    for season in temp_df['season'].unique():
        for matchweek in temp_df['round'].unique():
            temp = temp_df[(temp_df['season'] == season) & (temp_df['matchweek'] == matchweek)].copy()
            temp.sort_values(['pl_total_points', 'pl_total_goal_diff', 'team'], ascending=[False, False, True], inplace=True)
            temp['rank'] = range(1, len(temp) + 1)
            temp.loc[temp['rank'] > 20, 'rank'] = 20  # Replace ranks greater than 20 with 20
            ranks = pd.concat([ranks, temp])

    # Replace rank of first matchweek with NaN in the 'temp_df' dataframe
    ranks.loc[ranks['matchweek'] == 1, 'rank'] = np.nan

    # For each matchweek and each team, take the minimum rank
    ranks['rank'] = ranks.groupby(['season', 'matchweek', 'team'])['rank'].transform('min')
    
    # Merge back with the original dataframe
    matches = data.merge(ranks[['unique_match_id', 'rank']], on='unique_match_id', how='left')

    return matches

matches = calculate_league_position(matches)

matches = matches.rename(columns = {'rank': 'pl_position'})
matches['pl_position'] = matches['pl_position'].astype(float)

### Head to Head

# Create a sorted copy of the data
matches_sorted = matches.sort_values(['date']).copy()

# Create a set identifier for each match (sorted alphabetically to ensure that the same set represents the same pair of teams regardless of home/away status)
matches_sorted['set_identifier'] = matches_sorted['team'] + matches_sorted['opponent'] 

# Calculate last h2h result
matches_sorted['last_h2h'] = matches_sorted.groupby('set_identifier')['points'].transform(lambda x: x.shift())

# Calculate average points from last 5 head-to-head games
matches_sorted['last_h2h_form'] = matches_sorted.groupby('set_identifier')['points'].rolling(window=5, min_periods=1, closed = 'left').mean().reset_index(level=0, drop=True)

# Create a venue identifier (including the venue in the identifier)
matches_sorted['venue_identifier'] = matches_sorted['team'] + matches_sorted['opponent'] +  matches_sorted['venue']

# Calculate last h2h result at same venue
matches_sorted['last_h2h_venue'] = matches_sorted.groupby('venue_identifier')['points'].transform(lambda x: x.shift())

# Calculate average points from last 5 head-to-head games at the same venue
matches_sorted['last_h2h_venue_form'] = matches_sorted.groupby('venue_identifier')['points'].rolling(window=5, min_periods=1, closed = 'left').mean().reset_index(level=0, drop=True)

matches = matches_sorted.sort_values(['date']).copy()
columns_to_convert = ['last_h2h', 'last_h2h_form', 'last_h2h_venue', 'last_h2h_venue_form']
matches[columns_to_convert] = matches[columns_to_convert].astype(float)

### PL stats from previous season

# Sum all the points, goals scored, goals conceded, and calculate goal difference for each team in each season
season_stats = matches[matches['comp'] == 'Premier League'].groupby(['season', 'team']).agg({
    'points': 'sum',
    'gf': 'sum',
    'ga': 'sum'
}).reset_index()

# Rename the variables and increment the season by one
season_stats['season'] = season_stats['season'] + 1
season_stats = season_stats.rename(columns={'points': 'prev_season_points', 'gf': 'prev_season_gf', 'ga': 'prev_season_ga'})

# Merge with the match data
matches = matches.merge(season_stats, on=['season', 'team'], how='left')

# calculate previous season goal difference
matches['prev_season_goal_diff'] = matches['prev_season_gf'] - matches['prev_season_ga']
columns_to_convert = ['prev_season_points', 'prev_season_gf', 'prev_season_ga', 'prev_season_goal_diff']
matches[columns_to_convert] = matches[columns_to_convert].astype(float)

### Form over last 5 games

# Define a list of variables to calculate the form
variables = ['points', 'gf', 'ga', 'poss', 'xg', 'xga']

# Filter the dataframe to include only Premier League games
premier_league_matches = matches[matches['comp'] == 'Premier League'].copy()

# Calculate the form for each variable
for variable in variables:
    # Create a new column for the variable's form
    premier_league_matches[f'{variable}_pl_form'] = premier_league_matches.groupby('team')[variable].rolling(window=5, min_periods=1, closed='left').mean().reset_index(level=0, drop=True)


# create a list of form variables
form_variables =  [f'{variable}_pl_form' for variable in variables]
    
# Merge the form data into the original matches dataframe
matches = matches.merge(premier_league_matches[['unique_match_id'] + form_variables], on='unique_match_id', how='left')
columns_to_convert = form_variables
#['points_pl_form', 'gf_pl_form', 'ga_pl_form', 'poss_pl_form', 'xg_pl_form', 'xga_pl_form']
matches[columns_to_convert] = matches[columns_to_convert].astype(float)

### Dummy variable for promoted teams

# Create an empty column 'promoted' in the matches dataframe
matches['promoted'] = 0

# Get the unique seasons in the dataframe
unique_seasons = matches['season'].unique()

# Iterate over the unique seasons
for i, season in enumerate(unique_seasons):
    if i > 0: # ignore the first season
        prev_season_teams = matches[matches['season'] == unique_seasons[i - 1]]['team'].unique()
        current_season_teams = matches[matches['season'] == season]['team']
        
        # Update the 'promoted' column for teams not present in the previous season
        matches.loc[(matches['season'] == season) & (~current_season_teams.isin(prev_season_teams)), 'promoted'] = 1

### Make venue column binary

matches['home'] = np.where(matches['venue'] == 'Home', 1, 0)

### Merge with opponents PL position and promotion status

# Select the columns to calculate for the opponent
opponent_columns_ = ['date', 'team', 'opponent', 'days_since_last_game', 'games_played_last_21_days', 'pl_total_points', 'pl_total_gf',
                    'pl_total_ga', 'pl_total_goal_diff', 'pl_position', 'points_pl_form', 'gf_pl_form',
                    'ga_pl_form', 'poss_pl_form', 'xg_pl_form', 'xga_pl_form', 'prev_season_points',
                    'prev_season_gf', 'prev_season_ga', 'prev_season_goal_diff', 'promoted']


# Merge original data with data from the opponent
matches = matches.merge(matches[opponent_columns_], left_on=['date', 'team', 'opponent'], right_on=['date', 'opponent', 'team'],
                          suffixes=('', '_opponent'))

### Final clean and store

# filter to just contain premier league data
matches = matches[matches['comp'] == 'Premier League'].sort_values(['date'])

# collect only relevant variables for ML, i.e., nothing based on the actual game data as we won't have that information when predicting games
matches_final = matches[['unique_match_id','season','round','day','team','promoted','opponent','promoted_opponent','home','points',
                        'days_since_last_game','games_played_last_21_days','pl_total_points','pl_total_gf','pl_total_ga','pl_total_goal_diff','pl_position','last_h2h','last_h2h_form','last_h2h_venue','last_h2h_venue_form',
                        'prev_season_points','prev_season_gf','prev_season_ga','prev_season_goal_diff','points_pl_form','gf_pl_form','ga_pl_form','poss_pl_form','xg_pl_form','xga_pl_form',
                        'days_since_last_game_opponent','games_played_last_21_days_opponent','pl_total_points_opponent','pl_total_gf_opponent','pl_total_ga_opponent','pl_total_goal_diff_opponent','pl_position_opponent',
                         'points_pl_form_opponent','gf_pl_form_opponent', 'ga_pl_form_opponent', 'poss_pl_form_opponent','xg_pl_form_opponent', 'xga_pl_form_opponent','prev_season_points_opponent', 'prev_season_gf_opponent',
                         'prev_season_ga_opponent', 'prev_season_goal_diff_opponent']]

# remove the first premier league season from the data frame as we dont have enough data
matches_final = matches_final[matches_final['season'] != 1993].reset_index(drop=True)

# convert round to int
matches_final['round'] = matches_final['round'].astype(int)                                                                 

# save clean data in processed data file
matches_final.to_csv(f"{DATAPATH}/processed/matches_processed.csv", index=False)
