==> notebooks/03_Train_Test_Split.ipynb <==
<center><h1><font size=6> Splitting the Data into Training and Testing Sets </h1></center>

At this point in the process, I need to seperate out my training and testing datasets. I want to do this before I run any Exploratory Data Analysis (EDA) and advanced feature engineering/feature selection to keep the testing set completely unseen and prevent any possible data snooping bias, which could help me analyse patterns or behaviours in the test set.

But before splitting the data into test and train, I want to create two different models. This is because the extend of EPL data collection improved in the 2017-18 season and data on variables like expected goals started to be collected. These could be important features so I want to run a model using only more recent data which includes these variables and compare it to the simpler model which is trained on more historic data. 

### Load libraries and setup notebook configuration

# import packages
import pandas as pd 
import numpy as np
import os
from pathlib import Path
import warnings


# set pandas configurations
pd.set_option("display.precision", 2) # display to 1 decimpal place
pd.set_option("display.max.columns", None) # display all columns so we can view the whole dataset
pd.set_option('display.float_format', '{:.2f}'.format) # Disable scientific notation for pandas
warnings.filterwarnings("ignore", category=pd.errors.SettingWithCopyWarning) # Disable setting with copy warnings


# set directories
os.chdir('..') # change current working directory to the parent directory to help access files/directories at a higher level
DATAPATH = Path(r'data') # set data path


# import from source directory
from src import constants

### Load data from local file

# load EPL match data
matches = pd.read_csv(f"{DATAPATH}/processed/matches_processed.csv")
matches.head(5)

### Split full dataset into full and shorter data sets

# for the full dataset, drop the variables that are only recently collected
matches_full = matches.copy()
columns_to_drop = ['xg_pl_form', 'xga_pl_form', 'poss_pl_form', 'xg_pl_form_opponent', 'xga_pl_form_opponent', 'poss_pl_form_opponent']

matches_full = matches_full.drop(columns=columns_to_drop)

# data on expected goals came into play in the 2017-18 season
matches_short = matches[matches['season'] >= 2018]

### Split datasets into training and test sets

from sklearn.model_selection import train_test_split

# Specify the features and target variable
features_full = matches_full.drop('points', axis=1)
target_full = matches_full[['unique_match_id', 'points']]
features_short = matches_short.drop('points', axis=1)
target_short = matches_short[['unique_match_id', 'points']]

# Perform train-test split with stratified sampling of the target variable to ensure we have a representative sample of all results
X_train_full, X_test_full, y_train_full, y_test_full = train_test_split(features_full, target_full, test_size=0.2, stratify=target_full['points'], random_state=42)
X_train_short, X_test_short, y_train_short, y_test_short = train_test_split(features_short, target_short, test_size=0.2, stratify=target_short['points'], random_state=42)

# store in local data file
# Define the output directory
output_dir = f"{DATAPATH}/processed/"

# Save the train and test sets as CSV files
X_train_full.to_csv(os.path.join(output_dir, 'X_train_full.csv'), index=False)
X_test_full.to_csv(os.path.join(output_dir, 'X_test_full.csv'), index=False)
y_train_full.to_csv(os.path.join(output_dir, 'y_train_full.csv'), index=False)
y_test_full.to_csv(os.path.join(output_dir, 'y_test_full.csv'), index=False)

X_train_short.to_csv(os.path.join(output_dir, 'X_train_short.csv'), index=False)
X_test_short.to_csv(os.path.join(output_dir, 'X_test_short.csv'), index=False)
y_train_short.to_csv(os.path.join(output_dir, 'y_train_short.csv'), index=False)
y_test_short.to_csv(os.path.join(output_dir, 'y_test_short.csv'), index=False)
